Switch from js to ts
