+++ tools/csvToDatabase/package.json
@@ -10,6 +10,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/mysql": "^2.15.21",
     "mysql": "^2.18.1"
   }
 }
